requirements.txt
tensorflow==2.2.0
keras==2.3.1
numpy==1.17.5
scikit-learn==0.22.2.post1
pandas==1.1.5
protobuf==3.20.*
Django==3.1.*
pymysql==1.1.0
matplotlib==3.2.2
seaborn==0.10.1
django_compressor